Fix log folder ownership when restoring
When we take a backup from one server and restore on another one, the
log file is also backuped/restored, but the ffsync user uid on the origin server
might not match the uid on the new serveur. When it's not the case, the
user cannot write to this file, making the systemd unit fail.

This commit ensure all log files are owned by the new ffsync user.